refactor: reorganize gamification module into domain, application, and infrastructure layers.
