+++ backend/src/main/java/com/swipelab/gamification/application/GamificationOrchestrator.java
@@ -1,0 +1,6 @@
+package com.swipelab.gamification.application;
+
+public class GamificationOrchestrator {
+    // TODO: Need to implement
+
+}
