+++ backend/src/main/java/com/swipelab/gamification/domain/BadgeType.java
@@ -1,0 +1,5 @@
+package com.swipelab.gamification.domain;
+
+public class BadgeType {
+    
+}
