+++ backend/src/main/java/com/swipelab/gamification/domain/UserBadge.java
@@ -1,4 +1,4 @@
-package com.swipelab.model.entity;
+package com.swipelab.gamification.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
+
+import com.swipelab.model.entity.User;
 
 import java.time.LocalDateTime;
