+++ backend/src/main/java/com/swipelab/gamification/infrastructure/LeaderboardRepository.java
@@ -1,7 +1,8 @@
-package com.swipelab.repository;
+package com.swipelab.gamification.infrastructure;
 
-import com.swipelab.model.entity.Leaderboard;
+import com.swipelab.gamification.domain.Leaderboard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LeaderboardRepository extends JpaRepository<Leaderboard, Long> {
+
 }
