+++ backend/src/main/java/com/swipelab/gamification/infrastructure/UserBadgeRepository.java
@@ -1,11 +1,12 @@
-package com.swipelab.repository;
+package com.swipelab.gamification.infrastructure;
 
-import com.swipelab.model.entity.UserBadge;
+import com.swipelab.gamification.domain.UserBadge;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
+//Need to be changed
 @Repository
 public interface UserBadgeRepository extends JpaRepository<UserBadge, Long> {
     List<UserBadge> findByUser_Username(String username);
