--- backend/src/main/java/com/swipelab/model/enums/BadgeType.java
@@ -1,5 +1,0 @@
-package com.swipelab.model.enums;
-
-public class BadgeType {
-    
-}
