+++ backend/src/main/java/com/swipelab/service/ClassificationService.java
@@ -13,9 +13,9 @@ import com.swipelab.repository.ImageRepository;
 import com.swipelab.repository.LabelRepository;
 import com.swipelab.repository.UserRepository;
 import com.swipelab.service.classification.FraudDetectionService;
-import com.swipelab.service.gamification.BadgeService;
-import com.swipelab.service.gamification.PointsService;
-import com.swipelab.service.gamification.StreakService;
+import com.swipelab.gamification.application.BadgeService;
+import com.swipelab.gamification.application.PointsService;
+import com.swipelab.gamification.application.StreakService;
 import com.swipelab.service.user.CredibilityService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
